+++ jax/experimental/pallas/ops/gpu/attention.py
@@ -288,7 +288,7 @@ def mha(
       grid=grid_,
       in_specs=in_specs,
       out_specs=out_specs,
-      compiler_params=plgpu.TritonCompilerParams(
+      compiler_params=plgpu.CompilerParams(
           num_warps=num_warps_, num_stages=num_stages),
       out_shape=out_shape,
       debug=debug,
@@ -351,7 +351,7 @@ def _preprocess_backward(out, do, lse, block_q: int,
                        lambda i, j, k: (j, i, k, 0)),
       ],
       out_specs=pl.BlockSpec((None, None, block_q), lambda i, j, k: (j, k, i)),
-      compiler_params=plgpu.TritonCompilerParams(num_warps=4, num_stages=3),
+      compiler_params=plgpu.CompilerParams(num_warps=4, num_stages=3),
       out_shape=out_shape,
       debug=debug,
       interpret=interpret,
@@ -634,7 +634,7 @@ def _mha_backward(sm_scale: float, causal: bool, block_sizes: BlockSizes,
         name="mha_backward",
         debug=debug,
         interpret=interpret,
-        compiler_params=plgpu.TritonCompilerParams(
+        compiler_params=plgpu.CompilerParams(
             num_warps=num_warps_, num_stages=2
         ),
     )(q, k, v, segment_ids, out, do, lse, delta)